Refactor Register component: move from Pages to Screens and update imports
